Create basic scaffolding for new Controller
